+++ project/handballeague/src/main/java/com/handballleague/controllers/TeamController.java
@@ -1,10 +1,10 @@
 package com.handballleague.controllers;
 
 import com.handballleague.model.Team;
+import com.handballleague.services.JWTService;
 import com.handballleague.services.TeamService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,10 +14,12 @@ import java.util.List;
 @RequestMapping(path = "api/v1/teams")
 public class TeamController {
     private final TeamService teamService;
+    private final JWTService jwtService;
 
     @Autowired
-    public TeamController(TeamService teamService) {
+    public TeamController(TeamService teamService, JWTService jwtService) {
         this.teamService = teamService;
+        this.jwtService = jwtService;
     }
 
     @GetMapping
@@ -29,41 +31,55 @@ public class TeamController {
     }
 
     @PostMapping
-    public ResponseEntity<?> registerNewTeam(@Valid @RequestBody Team team) {
-
-        teamService.create(team);
-        return ResponseEntity.ok("Team created successfully");
-
+    public ResponseEntity<?> registerNewTeam(@Valid @RequestBody Team team, @RequestHeader(name = "Authorization") String token) {
+        ResponseEntity<?> response = jwtService.handleAuthorization(token, "admin");
+        if (response.getStatusCode().is2xxSuccessful()) {
+            teamService.create(team);
+            return ResponseEntity.ok("Team created successfully");
+        } else {
+            return response;
+        }
     }
 
     @PostMapping("/{teamId}/players")
-    public ResponseEntity<?> addPlayerToTeam(@PathVariable Long teamId, @RequestBody Long playerId) {
-        if (playerId == null || teamId == null) {
-            return ResponseEntity.badRequest().build();
+    public ResponseEntity<?> addPlayerToTeam(@PathVariable Long teamId, @RequestBody Long playerId, @RequestHeader(name = "Authorization") String token) {
+        ResponseEntity<?> response = jwtService.handleAuthorization(token, "admin");
+        ResponseEntity<?> response2 = jwtService.handleAuthorization(token, "captain");
+        if (response.getStatusCode().is2xxSuccessful() || response2.getStatusCode().is2xxSuccessful()) {
+            if (playerId == null || teamId == null) {
+                return ResponseEntity.badRequest().build();
+            }
+            Team updatedTeam = teamService.addPlayerToTeam(teamId, playerId);
+            return ResponseEntity.ok(updatedTeam);
+        } else {
+            return response2;
         }
-
-        Team updatedTeam = teamService.addPlayerToTeam(teamId, playerId);
-        return ResponseEntity.ok(updatedTeam);
-
     }
 
     @DeleteMapping("/{teamId}/players")
-    public ResponseEntity<?> removePlayerFromTeam(@PathVariable Long teamId, @RequestBody Long playerId) {
-        if (playerId == null || teamId == null) {
-            return ResponseEntity.badRequest().build();
+    public ResponseEntity<?> removePlayerFromTeam(@PathVariable Long teamId, @RequestBody Long playerId, @RequestHeader(name = "Authorization") String token) {
+        ResponseEntity<?> response = jwtService.handleAuthorization(token, "admin");
+        ResponseEntity<?> response2 = jwtService.handleAuthorization(token, "captain");
+        if (response.getStatusCode().is2xxSuccessful() || response2.getStatusCode().is2xxSuccessful()) {
+            if (playerId == null || teamId == null) {
+                return ResponseEntity.badRequest().build();
+            }
+            Team updatedTeam = teamService.removePlayerFromTeam(teamId, playerId);
+            return ResponseEntity.ok(updatedTeam);
+        } else {
+            return response2;
         }
-
-        Team updatedTeam = teamService.removePlayerFromTeam(teamId, playerId);
-        return ResponseEntity.ok(updatedTeam);
-
     }
 
     @DeleteMapping("/{teamId}")
-    public ResponseEntity<?> deleteTeam(@PathVariable("teamId") Long id) {
-
-        teamService.delete(id);
-        return ResponseEntity.ok("Team deleted successfully");
-
+    public ResponseEntity<?> deleteTeam(@PathVariable("teamId") Long id, @RequestHeader(name = "Authorization") String token) {
+        ResponseEntity<?> response = jwtService.handleAuthorization(token, "admin");
+        if (response.getStatusCode().is2xxSuccessful()) {
+            teamService.delete(id);
+            return ResponseEntity.ok("Team deleted successfully");
+        } else {
+            return response;
+        }
     }
 
     @GetMapping("/{teamId}")
@@ -75,10 +91,14 @@ public class TeamController {
     }
 
     @PutMapping("/{teamId}")
-    public ResponseEntity<?> updateTeam(@PathVariable Long teamId, @Valid @RequestBody Team team) {
-
-        Team newTeam = teamService.update(teamId, team);
-        return ResponseEntity.ok(newTeam);
-
+    public ResponseEntity<?> updateTeam(@PathVariable Long teamId, @Valid @RequestBody Team team, @RequestHeader(name = "Authorization") String token) {
+        ResponseEntity<?> response = jwtService.handleAuthorization(token, "admin");
+        ResponseEntity<?> response2 = jwtService.handleAuthorization(token, "captain");
+        if (response.getStatusCode().is2xxSuccessful() || response2.getStatusCode().is2xxSuccessful()) {
+            Team newTeam = teamService.update(teamId, team);
+            return ResponseEntity.ok(newTeam);
+        } else {
+            return response2;
+        }
     }
 }
